(PC-30040) refactor(typeguards): Move isString to new typeguard folder
