+++ src/shared/typeguards/isString.ts
